Move core enums and classes to domain package
