--- src/main/kotlin/no/nb/mlt/wls/core/data/Owner.kt
@@ -1,5 +1,0 @@
-package no.nb.mlt.wls.core.data
-
-enum class Owner {
-    NB
-}
