+++ src/main/kotlin/no/nb/mlt/wls/domain/Item.kt
@@ -1,10 +1,5 @@
 package no.nb.mlt.wls.domain
 
-import no.nb.mlt.wls.core.data.Environment
-import no.nb.mlt.wls.core.data.HostName
-import no.nb.mlt.wls.core.data.Owner
-import no.nb.mlt.wls.core.data.Packaging
-
 data class Item(
     val hostId: String,
     val hostName: HostName,
