+++ src/main/kotlin/no/nb/mlt/wls/domain/Owner.kt
@@ -1,0 +1,5 @@
+package no.nb.mlt.wls.domain
+
+enum class Owner {
+    NB
+}
