+++ src/main/kotlin/no/nb/mlt/wls/domain/Packaging.kt
@@ -1,4 +1,4 @@
-package no.nb.mlt.wls.core.data
+package no.nb.mlt.wls.domain
 
 enum class Packaging {
     BOX,
