+++ src/main/kotlin/no/nb/mlt/wls/domain/ports/outbound/ItemRepository.kt
@@ -1,6 +1,6 @@
 package no.nb.mlt.wls.domain.ports.outbound
 
-import no.nb.mlt.wls.core.data.HostName
+import no.nb.mlt.wls.domain.HostName
 import no.nb.mlt.wls.domain.Item
 import reactor.core.publisher.Mono
