+++ src/main/kotlin/no/nb/mlt/wls/infrastructure/synq/SynqAdapter.kt
@@ -1,7 +1,7 @@
 package no.nb.mlt.wls.infrastructure.synq
 
 import kotlinx.coroutines.reactor.awaitSingle
-import no.nb.mlt.wls.core.data.Packaging
+import no.nb.mlt.wls.domain.Packaging
 import no.nb.mlt.wls.domain.Item
 import no.nb.mlt.wls.domain.ports.outbound.StorageSystemFacade
 import no.nb.mlt.wls.infrastructure.synq.SynqError.Companion.createServerError
