+++ src/test/kotlin/no/nb/mlt/wls/order/controller/OrderControllerTest.kt
@@ -8,8 +8,8 @@ import kotlinx.coroutines.reactor.awaitSingleOrNull
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.test.runTest
 import no.nb.mlt.wls.EnableTestcontainers
-import no.nb.mlt.wls.core.data.HostName
-import no.nb.mlt.wls.core.data.Owner
+import no.nb.mlt.wls.domain.HostName
+import no.nb.mlt.wls.domain.Owner
 import no.nb.mlt.wls.infrastructure.synq.SynqError
 import no.nb.mlt.wls.order.model.Order
 import no.nb.mlt.wls.order.model.OrderLineStatus
