+++ src/test/kotlin/no/nb/mlt/wls/product/model/ItemModelConversionTest.kt
@@ -3,10 +3,10 @@ package no.nb.mlt.wls.product.model
 import no.nb.mlt.wls.application.restApi.ApiProductPayload
 import no.nb.mlt.wls.application.restApi.toApiPayload
 import no.nb.mlt.wls.application.restApi.toProduct
-import no.nb.mlt.wls.core.data.Environment
-import no.nb.mlt.wls.core.data.HostName
-import no.nb.mlt.wls.core.data.Owner
-import no.nb.mlt.wls.core.data.Packaging
+import no.nb.mlt.wls.domain.Environment
+import no.nb.mlt.wls.domain.HostName
+import no.nb.mlt.wls.domain.Owner
+import no.nb.mlt.wls.domain.Packaging
 import no.nb.mlt.wls.domain.Item
 import no.nb.mlt.wls.infrastructure.synq.SynqOwner
 import no.nb.mlt.wls.infrastructure.synq.SynqProductPayload
